requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14.0

# Database
psycopg2-binary>=2.9.9

# Excel parsing
pandas>=2.1.0
openpyxl>=3.1.2
xlrd>=2.0.1

# Jira integration
requests>=2.31.0

# Celery (optional)
celery>=5.3.0
redis>=5.0.0

# Environment variables
python-dotenv>=1.0.0

# Development
ipython>=8.17.0

